+++ app/models/core/search-model-array.js
@@ -17,7 +17,8 @@ var SearchModelArray = ModelArray.extend(Ember.SortableMixin, {
 	total_orders: readOnly('order'),
 	total_disputes: readOnly('dispute'),
 	total_transactions: Computed.sumAll('total_credits', 'total_debits', 'total_card_holds', 'total_refunds'),
-	total_funding_instruments: Computed.sumAll('total_bank_accounts', 'total_cards')
+	total_funding_instruments: Computed.sumAll('total_bank_accounts', 'total_cards'),
+	total_results: Computed.sumAll('total_orders', 'total_transactions', 'total_funding_instruments', 'total_customers')
 });
 
 export default SearchModelArray;