+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Updated search results view (#1489)
 * Added keyboard shortcuts to open and close the search modal
+* Bug fix for log search results
+* Display the most appropriate search results tab by default
 
 ### 1.1.1